missed scrcmd change and itemgfx updates

+++ include/battle_anim.h
@@ -40,7 +40,7 @@ struct BattleAnimBgData
 struct BattleAnimBackground
 {
     const u32 *image;
-    const u32 *palette;
+    const u16 *palette;
     const u32 *tilemap;
 };
