+++ include/pokemon_storage_system_internal.h
@@ -681,6 +681,6 @@ void SetBoxMonIconObjMode(u8 cursorPos, u8 objMode);
 void SetPartyMonIconObjMode(u8 cursorPos, u8 objMode);
 
 const u32 *GetItemIconPic(u16 itemId);
-const u32 *GetItemIconPalette(u16 itemId);
+const u16 *GetItemIconPalette(u16 itemId);
 
 #endif //GUARD_POKEMON_STORAGE_SYSTEM_INTERNAL_H
